chore: extract load-presence suite into src/suites/load-presence.ts
Part of REAL-1037.
