+++ test/e2e/realtime-check.ts
@@ -22,6 +22,7 @@ import { authorization } from "./src/suites/authorization.ts";
 import { loadBroadcastFromDb } from "./src/suites/load-broadcast-from-db.ts";
 import { loadBroadcastReplay } from "./src/suites/load-broadcast-replay.ts";
 import { connection } from "./src/suites/connection.ts";
+import { loadPresence } from "./src/suites/load-presence.ts";
 
 async function runLoadPostgresChangesTests(testUser: { email: string; password: string }) {
   suite("load-postgres-changes");
@@ -133,55 +134,6 @@ async function runLoadPostgresChangesTests(testUser: { email: string; password: 
       return measureThroughput(latencies, LOAD_MESSAGES, "DELETE events", LOAD_DELIVERY_SLO);
     } finally {
       await stopClient(supabase);
-    }
-  });
-}
-
-async function runLoadPresenceTests() {
-  suite("load-presence");
-
-  await sleep(RATE_LIMIT_PAUSE_MS);
-  await test("presence join throughput", async () => {
-    const CLIENTS = 10;
-    const observer = createClient(PROJECT_URL, ANON_KEY, { realtime: REALTIME_OPTS });
-    const senders: ReturnType<typeof createClient>[] = [];
-    try {
-      const topic = randomTopic();
-      const trackTimes = new Map<string, number>();
-      const latencies: number[] = [];
-
-      const observerChannel = observer
-        .channel(topic, { config: { broadcast: { self: true }, presence: { key: "observer" } } })
-        .on("presence", { event: "join" }, (e) => {
-          if (e.key === "observer") return;
-          const t = trackTimes.get(e.key);
-          if (t !== undefined) latencies.push(performance.now() - t);
-        });
-      await openChannel(observerChannel);
-
-      const clients = Array.from({ length: CLIENTS }, (_, i) => ({
-        client: createClient(PROJECT_URL, ANON_KEY, { realtime: REALTIME_OPTS }),
-        key: `client-${i}`,
-      }));
-      senders.push(...clients.map((c) => c.client));
-
-      const channels = await Promise.all(clients.map(async ({ client, key }) => {
-        const ch = client.channel(topic, { config: { presence: { key } } });
-        await openChannel(ch);
-        return { ch, key };
-      }));
-
-      await Promise.all(channels.map(({ ch, key }) => {
-        trackTimes.set(key, performance.now());
-        return ch.track({ key });
-      }));
-
-      await settle(() => latencies.length, CLIENTS, LOAD_SETTLE_MS);
-
-      return measureThroughput(latencies, CLIENTS, "presence joins", LOAD_DELIVERY_SLO);
-    } finally {
-      await Promise.all(senders.map((c) => stopClient(c)));
-      await stopClient(observer);
     }
   });
 }
@@ -1293,7 +1245,7 @@ async function runBroadcastReplayTests(_testUser: { email: string; password: str
 const descriptors: SuiteDescriptor[] = [
   connection,
   { name: "load-postgres-changes", label: "load-postgres-changes", needsDb: true, run: ({ testUser }) => runLoadPostgresChangesTests(testUser) },
-  { name: "load-presence", label: "load-presence", needsDb: false, run: () => runLoadPresenceTests() },
+  loadPresence,
   { name: "load-broadcast", label: "load-broadcast", needsDb: false, run: () => runLoadBroadcastTests() },
   loadBroadcastFromDb,
   loadBroadcastReplay,
